+++ test/mercado_pago_integration_test.dart
@@ -2,6 +2,18 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mercado_pago_integration/mercado_pago_integration.dart';
 
+final Map<String, Object> preference = {
+  'items': [
+    {
+      'title': 'Product',
+      'description': 'Description',
+      'quantity': 3,
+      'currency_id': 'ARS',
+      'unit_price': 1500,
+    }
+  ],
+  'payer': {'name': 'Buyer G.', 'email': 'test@test.com'},
+};
 void main() {
   const MethodChannel channel = MethodChannel('mercado_pago_integration');
 
@@ -20,7 +32,10 @@ void main() {
   test('getPlatformVersion', () async {
     expect(
         await MercadoPagoIntegration.startCheckout(
-            publicKey: "", checkoutPreferenceId: ""),
+          publicKey: "",
+          accessToken: "",
+          preference: preference,
+        ),
         'Starting');
   });
 }